# docs/openapi_flowers.yaml
openapi: 3.0.0
info: 
  title: "Цветы"
  version: 0.0.1
  description: |
    MVP из говна и палок 

    Для использования бота нужно сделать отдельную авторизацию с Oauth телеги. Пока стандартный шаблон.
    
    Last update: 00:00 27.02.24 by NF
servers:
  - url: http://localhost:8453/api/v1/
    description: API version 1
paths:
...
  /supplier/addProduct:
    post:
      summary: Добавление товара (+)
      tags:
        - Supplier
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CatalogItemCreation"
      responses:
        200:
          description: Responce with data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DefaultPing"
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /supplier/ordersForMe:
    get:
      summary: Заказы, которые сделали пользователи (+)
      tags:
        - Supplier
      security:
        - bearerAuth: []
      responses:
        200:
          description: Responce with data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OrderStatus"
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /supplier/setOrderStatus:
    post:
      summary: Установка статуса обработки заказа (+)
      tags:
        - Supplier
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: orderId
          required: true
          schema:
            $ref: "#/components/schemas/ProductId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                newStatus:
                  type: string
                  example: "Доставляется"
      responses:
        200:
          description: Responce with data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DefaultPing"
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /supplier/finishOrder:
    post:
      summary: Завершение заказа (+)
      tags:
        - Supplier
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: orderId
          required: true
          schema:
            $ref: "#/components/schemas/ProductId"
      responses:
        200:
          description: Responce with data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DefaultPing"
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /supplier/supplierReport:
    get:
      summary: Получение отчёта о продажах
      tags:
        - Supplier
      security:
        - bearerAuth: []
      responses:
        200:
          description: Responce with data
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    ID Заявки:
                      type: integer
                      example: 1
                    Дата создания (UTC):
                      type: string
                      example: "DD-MM-YYYY"
                    ID Товара:
                      type: integer
                      example: 1
                    Сумма:
                      type: integer
                      example: 1
                    Валюта:
                      type: string
                      example: "RUB"
                    Статус оплаты:
                      type: string
                      example: "Тест"
                    Комментарий банка:
                      type: string
                      example: "Тест"
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /supplier/myProductsList:
    get:
      summary: Просмотр моих товаров (+-)
      tags:
        - Supplier
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: start
          required: true
          schema:
            type: integer
            description: Номер первого продукта
            example: 0
        - in: query
          name: count
          required: false
          schema:
            type: integer
            description: Число продуктов в выдаче
            example: 20
        - in: query
          name: sort
          required: false
          schema:
            type: string
            description: |
             Один из вариантов: time_upscending(+), time_descending(+), rank_upscending(-), rank_descending(-), timesBought_upscending(-), timesBought_descending(-)
            example: default
      responses:
        200:
          description: Responce with data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CatalogItemsArray"
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    basicAuth:
      type: http
      scheme: basic
  schemas:
    FirstName:
      type: string
      description: Имя
      maxLength: 30
      pattern: "[a-zA-Z]+"
      example: Макака
    SecondName:
      type: string
      description: Фамилия
      maxLength: 50
      pattern: "[a-zA-Z]+"
      example: Тупая
    PhoneNumber:
      type: string
      description: Номер телефона
      maxLength: 16
      pattern: "/+[0-9]{11,15}"
      example: +78005553535
...
    OrderId:
      type: integer
      description: Номер заказа
      example: 1
    ProductId:
      type: integer
      description: Номер товара
      example: 1
    Comment:
      type: string
      description: Комментарий от заказчика
      maxLength: 300
      example: Домофон не работает, позвоните на номер телефона, указанный в заказе
    Adress:
      type: object
      description: Адрес гоя
      properties:
        Country:
          type: string
          description: "Country of costumer"
          example: "Russia"
        City:
          type: string
          description: "City/town or etc of costumer"
          example: "St. Petersburg"
        Street:
          type: string
          description: "Street or etc of costumer"
          example: "Lensoveta street"
        Building:
          type: string
          description: "Home number (with additional data like letter)"
          example: "23"
        Flat:
          type: string
          description: "Flat/Office number"
          example: "147"
...
    ProductIdArray:
      type: array
      description: массив номеров заказанных продуктов
      items:
        $ref: "#/components/schemas/ProductId"
...
    OrderStatus:
      type: object
      required:
        - orderId
        - orderStatus
        - orderCreatedTime
        - adress
        - comment
        - customerPhone
        - customerFirstName
        - customerSecondName
        - productIdArray
      properties:
        orderId:
          $ref: "#/components/schemas/OrderId"
        adress:
          $ref: "#/components/schemas/Adress"
        orderStatus:
          type: string
          description: Статус заказа # maybe we must use enum to create pretty bar (or add smth like stage number)
          example: Обрабатывается...
        orderCreatedTime:
          type: integer
          description: timestamp
          example: 1903093
        customerPhone:
          $ref: "#/components/schemas/PhoneNumber"
        customerFirstName:
          $ref: "#/components/schemas/FirstName"
        customerSecondName:
          $ref: "#/components/schemas/SecondName"
        comment:
          $ref: "#/components/schemas/Comment"  
        productIdArray:
          $ref: "#/components/schemas/ProductIdArray" 
...
    CatalogItemTitle:
      type: string
      description: "Краткое описание букета"
      example: "Букет из чего-то там"
    CatalogItemAuthor:
      type: string
      description: "Название магазина"
      example: "ООО Рога и копыта"
    CatalogItemImage:
      type: string
      description: "Ссылка на картинку с букетом"
      example: "https://example.com/test.png"
    CatalogItemCost:
      type: object
      properties:
        currency: # задел на будуюущее для расширения
          type: string
          description: "Валюта"
          example: RUB
        costNum:
          type: integer
          description: Число тубриков
          example: 30
    CatalogItemReviewsInfo:
      type: object
      properties:
        rank:
          type: number
          description: "Рейтинг товара c точностью до 1 знака после запятой с округлением вверх (4.001=4.1)"
          example: 4.8
        reviewsCount:
          type: integer
          description: Число отзыврв
          example: 30
    CatalogItem:
      type: object
      description: Элемент каталога
      properties:
        title:
          $ref: "#/components/schemas/CatalogItemTitle"
        author:
          $ref: "#/components/schemas/CatalogItemAuthor"
        image:
          $ref: "#/components/schemas/CatalogItemImage"
        productId:
          $ref: "#/components/schemas/ProductId"
        cost:
          $ref: "#/components/schemas/CatalogItemCost"
        reviews:
          $ref: "#/components/schemas/CatalogItemReviewsInfo"
    CatalogItemsArray:
      type: array
      description: Массив элементов каталога
      items:
        $ref: "#/components/schemas/CatalogItem"
    CatalogItemDescription:
      type: string
      description: Описание элемента каталога
      example: |
        Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book.
...
    CatalogItemCreation:
      type: object
      description: Описание товара со стороны продавца
      properties:
        title:
          $ref: "#/components/schemas/CatalogItemTitle"
        titleImage:
          $ref: "#/components/schemas/CatalogItemImage"
        additionalImages:
          type: array
          items:
            $ref: "#/components/schemas/CatalogItemImage"
        cost:
          $ref: "#/components/schemas/CatalogItemCost"
        description:
          $ref: "#/components/schemas/CatalogItemDescription"
    DefaultPing:
      type: object
      properties:
        data:
          type: string
          description: "DO NOT USE THIS FIELD! It's placeholder and can be removed anytime/"
          example: "gooooool!"
...
    Error:
      type: object
      required:
        - description
      properties:
        code:
          type: integer
          description: "Code of error (optional! it's better to get status code from responce directly!)"
          example: 500
        description:
          type: string
          description: "Description of error"
          example: description of error